+++ backfit_generator/CHANGELOG.md
@@ -1,3 +1,7 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.2] 30/08/2021
 
-* TODO: Describe initial release.
+* Now using PartBuilder instead of Shared Part Builder
+* \*Breaking Change\* - Use ".backfit.dart" for part files instead of ".g.dart"
+## [0.0.1]
+
+* Initial Release